pyproject.toml
[project]
name = "sql-review-agent"
version = "0.1.0"
description = "Multi-Agent Inner Loop DevOps for SQL Server — autonomous T-SQL review pipeline"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=1.0",
    "langchain-core>=1.0",
    "langchain-openai>=1.0",
    "langchain-anthropic>=1.0",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "sqlparse>=0.5",
    "azure-cosmos>=4.0",
    "azure-keyvault-secrets>=4.0",
    "azure-identity>=1.0",
    "opentelemetry-sdk>=1.0",
    "azure-monitor-opentelemetry>=1.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
